refactor : 객체의 책임 이동으로 응집도 높이기
