+++ src/main/kotlin/blackjack/BlackjackCardShuffle.kt
@@ -1,8 +1,19 @@
 package blackjack
 
 class BlackjackCardShuffle : CardShuffle {
+    private val cards: MutableList<Card> = mutableListOf()
+
+    init {
+        CardSymbol.values().forEach {
+            cards.add(Card(CardShape.SPADE, it))
+            cards.add(Card(CardShape.HEART, it))
+            cards.add(Card(CardShape.DIAMOND, it))
+            cards.add(Card(CardShape.CLOVER, it))
+        }
+    }
+
     override fun getCards(): List<Card> {
-        return Cards.getCards()
+        return cards
             .shuffled()
             .toList()
     }
