+++ src/main/kotlin/blackjack/Card.kt
@@ -2,5 +2,5 @@ package blackjack
 
 data class Card(val cardShape: CardShape, val cardSymbol: CardSymbol) {
 
-    fun show(): String = "%s%s".format(cardSymbol.cardNumber, cardShape.name)
+    override fun toString(): String = "%s%s".format(cardSymbol.cardNumber, cardShape.name)
 }
