+++ src/main/kotlin/blackjack/dto/ResultDto.kt
@@ -6,7 +6,7 @@ class ResultDto(private val players: List<UserRole>) {
 
     fun result() {
         players.forEach {
-            println("%s카드: %s - 결과: %s".format(it.name, it.cards.map { c -> c.show() }, it.getScore()))
+            println("%s카드: %s - 결과: %s".format(it.name, it.cards.map { c -> c.toString() }, it.getScore()))
         }
 
         println("최종 결과")
