+++ src/test/kotlin/blackjack/DeckTest.kt
@@ -6,12 +6,7 @@ import org.junit.jupiter.api.Test
 internal class DeckTest {
     @Test
     fun `덱을 처음 생성 시 카드를 52개 가진다`() {
-        val cardShuffle = object : CardShuffle {
-            override fun getCards(): List<Card> {
-                return Cards.getCards()
-            }
-        }
-        val deck = Deck(cardShuffle)
+        val deck = Deck(BlackjackCardShuffle())
         assertThat(deck.cards.size).isEqualTo(52)
     }
 
@@ -29,12 +24,7 @@ internal class DeckTest {
 
     @Test
     fun `카드를 가져온 후 남은 카드 갯 수 확인한다`() {
-        val cardShuffle = object : CardShuffle {
-            override fun getCards(): List<Card> {
-                return Cards.getCards()
-            }
-        }
-        val deck = Deck(cardShuffle)
+        val deck = Deck(BlackjackCardShuffle())
         deck.deal()
         assertThat(deck.cards.size).isEqualTo(51)
     }
